+++ packages/babel-preset-expo/__tests__/index-test.js
@@ -87,6 +87,32 @@ export default function App() {
     expect(code).toMatchSnapshot();
   });
 
+  it(`supports reanimated worklets`, () => {
+    expect(require.resolve('react-native-reanimated/plugin')).toBeDefined();
+
+    const samplesPath = path.resolve(__dirname, 'samples/worklet.js');
+
+    const options = {
+      babelrc: false,
+      presets: [[preset, { jsxRuntime: 'automatic' }]],
+      // Make the snapshot easier to read
+      retainLines: true,
+      caller,
+    };
+
+    const { code } = babel.transformFileSync(samplesPath, options);
+
+    expect(code).toMatchSnapshot();
+
+    expect(
+      babel.transformFileSync(samplesPath, {
+        ...options,
+        // Test that duplicate plugins make no difference
+        plugins: [require.resolve('react-native-reanimated/plugin')],
+      }).code
+    ).toBe(code);
+  });
+
   it(`supports classic JSX runtime`, () => {
     const options = {
       babelrc: false,