+++ packages/babel-preset-expo/__tests__/samples/worklet.js
@@ -1,0 +1,4 @@
+function someWorklet(greeting) {
+  'worklet';
+  console.log("Hey I'm running on the UI thread");
+}